+++ Content.Shared/_NF/Bank/Events/BalanceChangedMessage.cs
@@ -3,10 +3,10 @@ using Robust.Shared.Player;
 namespace Content.Shared._NF.Bank.Events;
 public sealed class BalanceChangedEvent : EntityEventArgs
 {
-    public readonly int Amount;
+    public readonly ulong Amount;
     public readonly ICommonSession Player;
 
-    public BalanceChangedEvent(int amount, ICommonSession player)
+    public BalanceChangedEvent(ulong amount, ICommonSession player)
     {
         Amount = amount;
         Player = player;
